SCRUM-58 Agregar la opción de cancelar y confirmar antes de eliminar el perfil

+++ app/src/main/java/es/unex/giiis/fitlife365/view/home/EditarPerfilFragment.kt
@@ -103,7 +103,11 @@ class EditarPerfilFragment : Fragment() {
 
         btnEliminar.setOnClickListener {
             if (user != null) {
-                eliminarUsuario(user)
+                mostrarDialogoConfirmacionEliminacion(user) { confirmed ->
+                    if (confirmed) {
+                        eliminarUsuario(user)
+                    }
+                }
             }
         }
 
@@ -158,6 +162,26 @@ class EditarPerfilFragment : Fragment() {
                 }
             }
         }
+    }
+    private fun mostrarDialogoConfirmacionEliminacion(user: User, callback: (Boolean) -> Unit) {
+        val builder = AlertDialog.Builder(requireContext())
+        builder.setTitle("Confirmación")
+        builder.setMessage("¿Estás seguro de eliminar tu perfil?")
+
+        // Configurar el botón de aceptar
+        builder.setPositiveButton("Aceptar") { _, _ ->
+            // El usuario ha confirmado, ejecutar el módulo eliminarUsuario
+            callback(true)
+        }
+
+        // Configurar el botón de cancelar
+        builder.setNegativeButton("Cancelar") { _, _ ->
+            // El usuario ha cancelado, no hacer nada
+            callback(false)
+        }
+
+        // Mostrar el cuadro de diálogo
+        builder.show()
     }
 
     private fun eliminarUsuario(user: User) {
